Title: 바이러스 검사, Time: 1603ms, Memory: 115MB, Status: Passed - CodeTree

+++ 240811/바이러스 검사/virus-detector.java
@@ -10,7 +10,7 @@ public class Main {
         int leader = Integer.parseInt(tokenizer.nextToken());
         int member = Integer.parseInt(tokenizer.nextToken());
 
-        int admins = n;
+        long admins = n;
         for(int i = 0; i < n; i++) {
             int customer = customers[i] - leader;
             if(customer > 0) {
